+++ dynaconf/typed/exceptions.py
@@ -1,0 +1,7 @@
+from dynaconf.validator import ValidationError
+
+
+class DynaconfSchemaError(Exception): ...
+
+
+__all__ = ["DynaconfSchemaError", "ValidationError"]
