+++ dynaconf/typed/utils.py
@@ -1,183 +1,18 @@
-"""Implements wrappers, types and helpers for Schema based settings."""
-
 from __future__ import annotations  # WARNING: remove this when debugging
 
-from collections.abc import Sequence
-from dataclasses import asdict
-from dataclasses import dataclass
-from pathlib import Path
 from typing import Annotated
-from typing import Any
 from typing import Callable
-from typing import cast
 from typing import get_args
 from typing import get_origin
 from typing import Union
 
-from dynaconf.base import LazySettings as OriginalDynaconf
 from dynaconf.base import Settings as BaseDynaconfSettings
-from dynaconf.utils.functional import Empty
 from dynaconf.utils.functional import empty
-from dynaconf.validator import ValidationError  # noqa
 from dynaconf.validator import Validator as BaseValidator
 
-
-class Validator:
-    """Define an interface to generate a valid BaseValidator.
-
-    The `operations` are::
-
-        eq: value == other
-        ne: value != other
-        gt: value > other
-        lt: value < other
-        gte: value >= other
-        lte: value <= other
-        is_type_of: isinstance(value, type)
-        is_in:  value in sequence
-        is_not_in: value not in sequence
-        identity: value is other
-        cont: contain value in
-        len_eq: len(value) == other
-        len_ne: len(value) != other
-        len_min: len(value) > other
-        len_max: len(value) < other
-        startswith: value.startswith(term)
-        endswith:  value.endswith(term)
-        condition: fn(value) is True
-
-
-    """
-
-    def __new__(
-        cls,
-        *,  # kw_only
-        # Operations
-        eq: Any = empty,
-        ne: Any = empty,
-        gt: Any = empty,
-        lt: Any = empty,
-        gte: Any = empty,
-        lte: Any = empty,
-        identity: Any = empty,
-        is_in: Any = empty,
-        is_not_in: Any = empty,
-        cont: Any = empty,
-        len_eq: Any = empty,
-        len_ne: Any = empty,
-        len_min: Any = empty,
-        len_max: Any = empty,
-        startswith: Any = empty,
-        endswith: Any = empty,
-        condition: Callable[[Any], bool] | None | Empty = empty,
-        # options
-        env: str | Sequence[str] | None | Empty = empty,
-        messages: dict[str, str] | None | Empty = empty,
-        cast: Callable[[Any], Any] | None | Empty = empty,
-        description: str | None | Empty = empty,
-        # To be implemented, a interface like this to generate a When Validator
-        # when: When | None = empty,
-    ):
-        kwargs = locals()
-        kwargs.pop("cls")
-        return BaseValidator(
-            **{k: v for k, v in kwargs.items() if v is not empty}
-        )
-
-
-@dataclass  # wanted to pass (kw_only=True) but that works only for 3.10+
-class Options:
-    """Options to configure the initialization of Dynaconf Object.
-
-    NOTE: This must be used only with typed implementation of Dynaconf.
-
-    All options are just bypassed to the Dynaconf `__init__` method,
-    except those starting with `_`.
-
-    Options starting with `_` are only used on `__new__` when building the
-    new Dynaconf object, those values assumes opinionated defaults and rarely
-    needs to be changed.
-    """
-
-    # NOTE: On a dataclass the sentinel must be a Type not an instance
-
-    env: str | type = Empty
-    environments: bool | type = Empty
-    envvar_prefix: str | type = Empty
-    root_path: str | Path | type = Empty
-    settings_files: list[str] | type = Empty
-    settings_file: str | type = Empty
-
-    _trigger_validation: bool = True
-
-    def as_dict(self):
-        options_dict = asdict(self)
-        options_dict = {
-            k: v
-            for k, v in options_dict.items()
-            if (not k.startswith("_")) and v is not Empty
-        }
-        return options_dict
-
-
-class DynaconfSchemaError(Exception): ...
-
-
-class DictValue:
-    """A dictvalue Subtype is actually a Dict, so instantiating it just returns
-    the dict."""
-
-    def __new__(cls, **kwargs):
-        defaults = {}
-        schema_annotations = get_annotations(cls)
-        for name, annotation in schema_annotations.items():
-            value = getattr(cls, name, empty)
-            if value is not empty:
-                defaults[name] = value
-        defaults.update(kwargs)
-        return defaults
-
-
-class Dynaconf(BaseDynaconfSettings):
-    """Implementation of Dynaconf that is aware of type annotations."""
-
-    def __new__(cls, *args, **kwargs):
-        options = getattr(cls, "dynaconf_options", Options())
-        defaults, validators = extract_defaults_and_validators_from_typing(cls)
-        passed_in_validators = kwargs.pop("validators", [])
-        validators += passed_in_validators
-
-        # Set init options for Dynaconf coming first from Options on Schema
-        init_options = options.as_dict()
-        # Add defaults defined on class as it was passed to Dynaconf(k=v)
-        # It will take first the default from cls, then from DictValue type.
-        init_options.update(defaults)
-        # Override from kwargs as if matching key was passed to Settings(k=v)
-        init_options.update(kwargs)
-
-        # Instantiate the new Dynaconf instance
-        new_cls = OriginalDynaconf(*args, **init_options)
-        new_cls.__annotations__ = cls.__annotations__
-
-        if validators:
-            new_cls.validators.register(*validators)
-
-        # Trigger validation if not explicitly disabled
-        if options._trigger_validation:
-            new_cls.validators.validate_all()
-
-        return cast(cls, new_cls)
-
-
-def is_union(annotation) -> bool:
-    """Tell if an annotation is strictly Union."""
-    return get_origin(annotation) is Union
-
-
-def is_optional(annotation) -> bool:
-    """Tell if an annotation is strictly typing.Optional or Union[**, None]"""
-    return is_union(annotation) and type(None) in get_args(annotation)
-
+from .compat import get_annotations
+from .exceptions import DynaconfSchemaError
+from .types import DictValue
 
 # Types that can be used on `field: list[x]`
 ALLOWED_ENCLOSED_TYPES = (
@@ -191,32 +26,6 @@ ALLOWED_ENCLOSED_TYPES = (
     tuple,
     dict,
 )
-
-
-def get_annotations(schema_cls) -> dict:
-    """NOTE: move to inspect.get_annotations when Python 3.9 is dropped"""
-    if isinstance(schema_cls, type):
-        return schema_cls.__dict__.get("__annotations__", {})
-    else:
-        return getattr(schema_cls, "__annotations__", {})
-
-
-def get_type_name(_type):
-    if name := getattr(_type, "__name__", None):
-        return name
-    return str(_type)
-
-
-def get_type_and_args(annotation) -> tuple:
-    """From annotation name: T[T, args] extract type, args, validators"""
-    _annotated_validators = None
-    _type_args: tuple = tuple()
-    _type = get_origin(annotation)  # type from X[type,]
-    if _type is Annotated:
-        _type, *_annotated_validators = get_args(annotation)
-    else:
-        _type_args = get_args(annotation)
-    return _type, _type_args, _annotated_validators
 
 
 def extract_defaults_and_validators_from_typing(
@@ -337,6 +146,34 @@ def extract_defaults_and_validators_from_typing(
             validators.append(_validator)
 
     return defaults, validators
+
+
+def get_type_name(_type):
+    if name := getattr(_type, "__name__", None):
+        return name
+    return str(_type)
+
+
+def get_type_and_args(annotation) -> tuple:
+    """From annotation name: T[T, args] extract type, args, validators"""
+    _annotated_validators = None
+    _type_args: tuple = tuple()
+    _type = get_origin(annotation)  # type from X[type,]
+    if _type is Annotated:
+        _type, *_annotated_validators = get_args(annotation)
+    else:
+        _type_args = get_args(annotation)
+    return _type, _type_args, _annotated_validators
+
+
+def is_union(annotation) -> bool:
+    """Tell if an annotation is strictly Union."""
+    return get_origin(annotation) is Union
+
+
+def is_optional(annotation) -> bool:
+    """Tell if an annotation is strictly typing.Optional or Union[**, None]"""
+    return is_union(annotation) and type(None) in get_args(annotation)
 
 
 def validator_condition_factory(validators, prefix, _type) -> Callable:
